// types.ts
export interface SuccessStory {
  id: string;
  name: string;
  discipline: string;
  image: string;
  quote: string;
  story: string;
}

export interface BenefitEvent {
  id: string;
  title: string;
  artist: string;
  date: string;
  location: string;
  price: string;
  image: string;
}

export interface Partner {
  name: string;
  logo: string;
}